scraper moved to the fast-api folder
